WIP - Module system and eventually adding yup validation
Moving all sub modules to a standard system and will be changing validation over to yup
